requirements.txt
aiogram==3.22.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
openai==1.55.3
